fix: multi-resource delete
